feature: implement user registration with unique-email detection

+++ include/auth/authService.h
@@ -27,6 +27,6 @@ struct RegisterError {
  * @param password The new user's plaintext password.
  * @return The new user's id on success, or a RegisterError on failure.
  */
-std::expected<int, RegisterError> registerUser(SqliteDb& db, const std::string& email, const std::string& password);
+std::expected<int64_t, RegisterError> registerUser(SqliteDb& db, const std::string& email, const std::string& password);
 
 }
